Photos can be called as dataframe (at will)
This update contains the streamline of the photos API to be viewable as a pandas dataframe. From this, we will be able to put this into a JSON format (or any format for that manner) to connect to streamlit. For now , for ease, simplicity, and accessibility its as a pandas.dataframe